# openapi.yml
openapi: 3.0.0
info:
  title: Promocode Service API
  version: 1.0.0
paths:
  /create_promocode:
    post:
      summary: Create a new promocode
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        201:
          description: Promocode created successfully
        400:
          description: Invalid data
        401:
          description: Unauthorized

  /delete_promocode/{promocode_id}:
    delete:
      summary: Delete a promocode by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: promocode_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Promocode deleted successfully
        401:
          description: Unauthorized
        404:
          description: Promocode not found

  /update_promocode/{promocode_id}:
    put:
      summary: Update a promocode by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: promocode_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        200:
          description: Promocode updated successfully
        400:
          description: Invalid data
        401:
          description: Unauthorized
        404:
          description: Promocode not found

  /get_promocode/{promocode_id}:
    get:
      summary: Get a promocode by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: promocode_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Promocode found
        401:
          description: Unauthorized
        404:
          description: Promocode not found

  /list_promocodes:
    get:
      summary: Get a list of all promocodes
      security:
        - bearerAuth: []
      responses:
        200:
          description: A list of promocodes
        401:
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT